backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
faiss-cpu
numpy
PyPDF2
